% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.R
\name{bal}
\alias{bal}
\title{Balancing by Either Inputs or Outputs}
\usage{
bal(T.star = "matrix", method = c("input", "output"))
}
\arguments{
\item{T.star}{Extended, unbalanced matrix.}

\item{method}{Balance by inputs or outputs.}
}
\value{
Returns an extended matrix for balancing by inputs or outputs.
}
\description{
Dependency for the \code{balance} function.
}
\references{
Fath, B.D. and S.R. Borrett. 2006. A MATLAB function for network
environ analysis. Environmental Modelling & Software 21:375-405.
}
\seealso{
\code{\link{balance}}
}
\author{
Matthew K. Lau Stuart R. Borrett
}
