% Please edit documentation in R/enaR-package.R
\docType{data}
\name{bgcModels}
\alias{bgcModels}
\title{Bigeochemical Cycling Models}
\description{
A set of 43 biogeochemical cycling models compiled by the SEE Lab at UNCW.
}
\references{
Borrett, S. R., and M. K. Lau. In Prep. enaR: An R package for
Ecological Network Analysis. Ecological Modeling and Software.
}
\keyword{datasets}
