% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaFlow.R
\name{enaFlow}
\alias{enaFlow}
\title{enaFlow --- flow analysis
INPUT = network object
OUTPUT = list of flow statistics}
\usage{
enaFlow(x, zero.na = TRUE, balance.override = FALSE)
}
\arguments{
\item{x}{a network object.  This includes all weighted flows into and out of
each node.}

\item{zero.na}{LOGICAL: should NA values be converted to zeros.}

\item{balance.override}{Flow analysis assumes the network model is at
steady-state (inputs = outputs).  Setting balance.override = TRUE allows the
function to be run on unbalanced models.}
}
\description{
M. Lau | July 2011
---------------------------------------------------
enaFlow --- flow analysis INPUT = network object OUTPUT = list of flow
statistics
}
\details{
M. Lau | July 2011 ---------------------------------------------------
enaFlow --- flow analysis INPUT = network object OUTPUT = list of flow
statistics

M. Lau | July 2011 --------------------------------------------------- Flow
Analyses of Ecological Networks

Performs the primary throughflow analysis developed for input-output
systems.  It returns a vector of throughflows, the input and output oriented
matrices for "direct flow intensities" and "integral flow intensities", and
a set of flow based network statistics.
}
\examples{



data(troModels)
F = enaFlow(troModels[[6]])  # completes the full analysis
F$ns  # returns just the network statisics



}
\references{
Borrett, S. R., Freeze, M. A., 2011. Reconnecting environs to
their environment. Ecol. Model. 222, 2393-2403.

Fath, B. D., Borrett, S. R. 2006. A Matlab function for Network Environ
Analysis.  Environ. Model. Softw. 21, 375-405.

Fath, B. D., Patten, B. C., 1999. Review of the foundations of network
environ analysis. Ecosystems 2, 167-179.

Finn, J. T., 1976. Measures of ecosystem structure and function derived from
analysis of flows. J. Theor. Biol. 56, 363-380.

Patten, B.C. Higashi, M., Burns, T. P. 1990. Trophic dynamics in ecosystem
networks: significance of cycles and storage.  Ecol. Model. 51, 1-28.

Schramski, J. R., Kazanci, C., Tollner, E. W., 2011. Network environ theory,
simulation and EcoNet 2.0. Environ. Model. Softw. 26, 419-428.

Szyrmer, J., Ulanowicz, R. E., 1987. "Total Flows in Ecosystems". Ecol. Mod. 35:123-136.

Ulanowicz, R.E., 2004. Quantitative methods for ecological network analysis.
Comput. Biol. Chem. 28, 321-33

Ulanowicz, R.E., Holt, R.D., Barfield, M., 2014. Limits on ecosystem trophic
complexity: insights from ecological network analysis.  Ecology Letters
17:127-136.
}
\seealso{
\code{\link{read.scor},\link{read.wand},\link{enaStorage},\link{enaUtility}}
}
\author{
Matthew K. Lau Stuart R. Borrett
}
