% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaStorage.R
\name{enaStorage}
\alias{enaStorage}
\title{Storage Analyses of Ecological Networks}
\usage{
enaStorage(x, balance.override = FALSE)
}
\arguments{
\item{x}{A network object.  This This includes all weighted flows into and
out of each vertex as well as the amount of energy--matter stored at each
vertex.}

\item{balance.override}{LOGICAL: should an imbalanced model be analyzed?  If
FALSE, the functions checks to make sure the network model provided is at
steady-state.  If TRUE, then the function will run without ensuring that the
model meets the steady-state assumption.}
}
\description{
Calculates storage-based Ecological Network Analyses.
}
\examples{
data(oyster)
S <- enaStorage(oyster)
attributes(S)
}
\references{
Barber, M. C. 1978a. A Markovian Model for Ecosystem Flow Analysis. Ecol. Model. 5(3):193-206.

Barber, M. C. 1978b. A Retrospective Markovian Model for Ecosystem Resource Flow. Ecol. Model. 5(2): 125-35.

Barber, M. C. 1979. A Note Concerning Time Parameterization of Markovian Models of Ecosystem Flow Analysis. Ecol. Model. 6(4): 323-28.

Matis, J. H., Patten, B. C. 1981. Environ analysis of linear
compartmental systems: the static, time invariant case.  Bulletin of the
International Statistical Institute, 48: 527-565.

Fath, B. D., Patten, B. C. 1999.  Review of the foundations of network
enviorn analysis.  Ecosystems 2:167-179.

Fath, B. D. Patten, B. C., Choi, J. 2001.  Compementarity of ecological goal
functions.  Journal of Theoretical Biology 208: 493-506.

Fath, B. D., Borrett, S. R. 2006. A MATLAB function for Network Environ
Analysis.  Environmental Modelling & Software 21:375-405.
}
\seealso{
\code{\link{read.scor},\link{read.wand},\link{enaFlow},\link{enaUtility}}
}
\author{
Matthew K. Lau Stuart R. Borrett
}
\keyword{enaFlow}
\keyword{read.scor}
