% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orient.R
\name{get.orient}
\alias{get.orient}
\title{Returns the global orientation}
\usage{
get.orient()
}
\description{
Returns the current setting for the expected orientation of all
matrices, which is either 'rc' (DEFAULT) or 'school' (output
orientation as expected for the school of analysis for a given
function). This function is intended to provide increase flexibility for users of both the Patten and Ulanowicz schools of ENA.
}
\author{
M.K. Lau and S.R. Borrett
}
