% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keystoneness.r
\name{keystoneness}
\alias{keystoneness}
\title{keystoneness  INPUT = network object OUTPUT = keystoneness vector}
\usage{
keystoneness(x = "model")
}
\arguments{
\item{x}{a network object.}
}
\value{
\item{ks}{vector of the keystoneness values}
}
\description{
Calculates the level of keystoneness for each species (node) in the
network following Libralato et al. 2006. The assumption is that the
input network model represents a food web.
}
\examples{

data(troModel)
ks <- keystoneness(troModels[[38]])
show(ks)

}
\references{
Libralato, S, Christensen, V., Pauly, D. 2006. A method for
identifying keystone speices in food web models. Ecol
Model. 195:153-171.
}
\seealso{
\code{\link{enaMTI}}
}
\author{
Stuart R. Borrett
}
