% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTrophicLevel.r
\name{meanTrophicLevel}
\alias{meanTrophicLevel}
\title{meanTrophicLevel
INPUT = network data object, min trophic level choice
OUTPUT = mean trophic level}
\usage{
meanTrophicLevel(x = "model", minTL = 2)
}
\arguments{
\item{x}{a network object.  This includes all weighted flows into and out of
each node.  The model needs to be at steady-state}

\item{minTL}{is the minimum trophic level to be included in the calculation fo the mean trophic level.   The default is 2, but users often use 3.25 and 4.}
}
\description{
==============================
}
\examples{

data(enaModels)
meanTrophicLevel(enaModels[[12]], minTL = 3.25)
}
\author{
Stuart R. Borrett
}
