% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaR-package.R
\docType{data}
\name{oyster}
\alias{oyster}
\title{Intertidal Oyster Reef Ecosystem Model}
\description{
Intertidal oyster reef ecosystem model created by Dame and Patten (1981).
Data were taken from Patten (1985).  Model flows are in kcal m^-2 day^-1;
storage data is kcal m^-2.
}
\references{
Dame, R. F., and B. C. Patten. 1981. Analysis of energy flows in
an intertidal oyster reef. Marine Ecology Progress Series 5:115-124.

Patten, B. C. 1985. Energy cycling, length of food chains, and direct versus
indirect effects in ecosystems. Can. Bull. Fish. Aqu. Sci. 213:119-138.
}
\keyword{datasets}

