% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lindeman.R
\name{plot.lindeman}
\alias{plot.lindeman}
\title{plot.lindeman}
\usage{
\method{plot}{lindeman}(x = "model", enatroagg = "troagg", primprod = NULL, type = 1)
}
\arguments{
\item{x}{an ENA network object.}

\item{enatroagg}{the results of the enaTroAgg function applied to the model}

\item{primprod}{a vector of the nodes that are primary producers}

\item{type}{switches between two types of plots: 1 = I and D are separate, 2 = I and D are combined}
}
\description{
INPUT = network object
OUTPUT = plot of the lindeman spine

Dr. Ulrike Schückel, Landesbetrieb für Küstenschutz, Nationalpark
und Meeresschutz Schleswig-Holstein, Nationalparkverwaltung,
Schlossgarten 1, 25832 Tönning

Applies the enaTroAgg function and creates a plot of the Lindeman Spine
}
\examples{
data(enaModels)
model <- enaModels[[8]]
enatroagg <- enaTroAgg(model)
plot.lindeman(model,enatroagg, type = 1)

}
\references{
Style of the plot according to Baird et al., 2004, 2007
}
\seealso{
\code{\link{enaTroAgg}}
}
\author{
Ulrike Schuckel,  Stuart R. Borrett
}
