% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EcoNet.R
\name{read.EcoNet}
\alias{read.EcoNet}
\title{### MKLau 1Mar2016
Read an EcoNet model.}
\usage{
read.EcoNet(file = "file path", verbose = FALSE, parse = FALSE)
}
\arguments{
\item{file}{Path to an EcoNet formatted file}

\item{verbose}{LOGICAL: should warnings be suppressed?}

\item{parse}{Is the file input parsed from a webpage?}
}
\value{
Returns the model formatted as a network object.
}
\description{
This function allows the user to access models that are formatted for
EcoNet, the web-based interface for conducting ENA
(http://eco.engr.uga.edu/), by Caner Kazanci at the University of Georgia.
}
\author{
Matthew K. Lau
}
\references{
Kazanci, C., 2007. EcoNet: A new software for ecological
modeling, simulation and network analysis, Ecol. Model., Vol 208/1 pp 3-8.
}
\seealso{
\code{\link{EcoNetWeb}}
}

