% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scifix.R
\name{scifix}
\alias{scifix}
\title{scifix Corrects missing e or E in scientific notation}
\usage{
scifix(x)
}
\arguments{
\item{x}{A numeric or character scalar.}
}
\value{
Returns a numeric scalar in appropriate scientific notation.
}
\description{
This is a support function that corrects the scientific notation in SCOR
formatted data files.
}
\seealso{
\code{\link{read.scor}}
}
\author{
Matthew K. Lau
}
