% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaR-package.R
\docType{data}
\name{troModels}
\alias{troModels}
\title{Trophic Models}
\description{
A set of 58 trophic models compiled by the SEE Lab at UNCW.
}
\references{
Borrett, S. R., and M. K. Lau. In Prep. enaR: An R package for
Ecological Network Analysis. Ecological Modeling and Software.
}
\keyword{datasets}

