\name{bcratio}
\alias{bcratio}
\title{
  Calculates the ratio of positive to negative elements in a network.
}
\description{
  Dependent function for the enaUtility function.
}
\usage{
bcratio(x = "matrix")
}
\arguments{
  \item{x}{
    A matrix of flow values.
  }
}

\value{
  Returns the ratio of positive to negative elements in the flow matrix.
}
\references{
  Fath, B.D. and S.R. Borrett. 2006. A MATLAB function for network
  environ analysis. Environmental Modelling & Software 21:375-405.
}
\author{
  Stuart R. Borrett (borretts@uncw.edu)
}

\seealso{
  \code{\link{enaUtility}}
}

