% Please edit documentation in R/eigenCentrality.R
\name{eigenCentrality}
\alias{eigenCentrality}
\title{Calculates the Eigen Centrality of a Network}
\usage{
eigenCentrality(x = "matrix")
}
\arguments{
\item{x}{A matrix defining a network graph.}
}
\value{
Returns the eigen based centrality of the network.
}
\description{
Calculates the eigen centrality of a network.
}
\author{
Stuart R. Borrett Matthew K. Lau
}
\references{
Bonacich, P., 1987. Power and centrality: a family of measures.
American Journal of Sociology 92: 1170-1182.
}
