\name{enaModelInfo}
\alias{enaModelInfo}
\docType{data}
\title{
  Ecosystem Model Information
}
\description{
Model information for the set of ecosystem models compiled by the SEE Lab at UNCW.
}

\references{
  Borrett, S. R., and M. K. Lau. In Prep. enaR: An R package for
  Ecological Network Analysis. Ecological Modeling and Software.
}
\usage{data(enaModelInfo)}


\keyword{datasets}
