% Please edit documentation in R/enaR-package.R
\docType{data}
\name{m.list}
\alias{m.list}
\title{Sub-set of the Larger Ecosystem Models}
\description{
A set of ecosystem models compiled by the SEE Lab at UNCW.
}
\references{
Borrett, S. R., and M. K. Lau. In Prep. enaR: An R package for
Ecological Network Analysis. Ecological Modeling and Software.
}
\keyword{datasets}
