\name{scifix}
\alias{scifix}
\title{
  Standardizes scientific notation.
}
\description{
  This is a support function that corrects the scientific notation in SCOR formatted
  data files.
}
\usage{
scifix(x)
}
\arguments{
  \item{x}{
    A numeric or character scalar.
  }
}

\value{
  Returns a numeric scalar in appropriate scientific notation.
}

\author{
  Matthew K. Lau (mkl48@nau.edu)
}

\seealso{
  \code{\link{read.scor}}
}

