\name{troModels}
\alias{troModels}
\docType{data}
\title{
Trophic Models
}
\description{
A set of 58 trophic models compiled by the SEE Lab at UNCW.
}

\references{
Borrett, S. R., and M. K. Lau. In Prep. enaR: An R package for
Ecological Network Analysis. Ecological Modeling and Software.

}
\usage{data(troModels)}


\keyword{datasets}
